+++ cli/internal/cloudcmd/create_test.go
@@ -14,7 +14,6 @@ import (
 	"testing"
 
 	"github.com/edgelesssys/constellation/v2/internal/cloud/cloudprovider"
-	"github.com/edgelesssys/constellation/v2/internal/cloud/cloudtypes"
 	"github.com/edgelesssys/constellation/v2/internal/config"
 	"github.com/edgelesssys/constellation/v2/internal/state"
 	"github.com/stretchr/testify/assert"
@@ -33,36 +32,17 @@ func TestCreator(t *testing.T) {
 		LoadBalancerIP: "192.0.2.1",
 	}
 
-	wantAzureState := state.ConstellationState{
-		CloudProvider: cloudprovider.Azure.String(),
-		AzureControlPlaneInstances: cloudtypes.Instances{
-			"id-0": {PrivateIP: "192.0.2.1", PublicIP: "192.0.2.1"},
-			"id-1": {PrivateIP: "192.0.2.1", PublicIP: "192.0.2.1"},
-		},
-		AzureWorkerInstances: cloudtypes.Instances{
-			"id-0": {PrivateIP: "192.0.2.1", PublicIP: "192.0.2.1"},
-			"id-1": {PrivateIP: "192.0.2.1", PublicIP: "192.0.2.1"},
-			"id-2": {PrivateIP: "192.0.2.1", PublicIP: "192.0.2.1"},
-		},
-		AzureSubnet:               "subnet",
-		AzureNetworkSecurityGroup: "network-security-group",
-		AzureWorkerScaleSet:       "workers-scale-set",
-		AzureControlPlaneScaleSet: "controlplanes-scale-set",
-	}
-
 	someErr := errors.New("failed")
 
 	testCases := map[string]struct {
-		tfClient          terraformClient
-		newTfClientErr    error
-		azureclient       azureclient
-		newAzureClientErr error
-		libvirt           *stubLibvirtRunner
-		provider          cloudprovider.Provider
-		config            *config.Config
-		wantState         state.ConstellationState
-		wantErr           bool
-		wantRollback      bool // Use only together with stubClients.
+		tfClient       terraformClient
+		newTfClientErr error
+		libvirt        *stubLibvirtRunner
+		provider       cloudprovider.Provider
+		config         *config.Config
+		wantState      state.ConstellationState
+		wantErr        bool
+		wantRollback   bool // Use only together with stubClients.
 	}{
 		"gcp": {
 			tfClient:  &stubTerraformClient{state: wantGCPState},
@@ -114,39 +94,6 @@ func TestCreator(t *testing.T) {
 			wantErr:      true,
 			wantRollback: !failOnNonAMD64,
 		},
-		"azure": {
-			azureclient: &fakeAzureClient{},
-			provider:    cloudprovider.Azure,
-			config:      config.Default(),
-			wantState:   wantAzureState,
-		},
-		"azure newAzureClient error": {
-			newAzureClientErr: someErr,
-			provider:          cloudprovider.Azure,
-			config:            config.Default(),
-			wantErr:           true,
-		},
-		"azure CreateVirtualNetwork error": {
-			azureclient:  &stubAzureClient{createVirtualNetworkErr: someErr},
-			provider:     cloudprovider.Azure,
-			config:       config.Default(),
-			wantErr:      true,
-			wantRollback: true,
-		},
-		"azure CreateSecurityGroup error": {
-			azureclient:  &stubAzureClient{createSecurityGroupErr: someErr},
-			provider:     cloudprovider.Azure,
-			config:       config.Default(),
-			wantErr:      true,
-			wantRollback: true,
-		},
-		"azure CreateInstances error": {
-			azureclient:  &stubAzureClient{createInstancesErr: someErr},
-			provider:     cloudprovider.Azure,
-			config:       config.Default(),
-			wantErr:      true,
-			wantRollback: true,
-		},
 		"unknown provider": {
 			provider: cloudprovider.Unknown,
 			config:   config.Default(),
@@ -163,9 +110,6 @@ func TestCreator(t *testing.T) {
 				newTerraformClient: func(ctx context.Context, provider cloudprovider.Provider) (terraformClient, error) {
 					return tc.tfClient, tc.newTfClientErr
 				},
-				newAzureClient: func(subscriptionID, tenantID, name, location, resourceGroup string) (azureclient, error) {
-					return tc.azureclient, tc.newAzureClientErr
-				},
 				newLibvirtRunner: func() libvirtRunner {
 					return tc.libvirt
 				},
@@ -176,19 +120,11 @@ func TestCreator(t *testing.T) {
 			if tc.wantErr {
 				assert.Error(err)
 				if tc.wantRollback {
-					switch tc.provider {
-					case cloudprovider.QEMU:
-						cl := tc.tfClient.(*stubTerraformClient)
-						assert.True(cl.destroyClusterCalled)
-						assert.True(cl.cleanUpWorkspaceCalled)
+					cl := tc.tfClient.(*stubTerraformClient)
+					assert.True(cl.destroyClusterCalled)
+					assert.True(cl.cleanUpWorkspaceCalled)
+					if tc.provider == cloudprovider.QEMU {
 						assert.True(tc.libvirt.stopCalled)
-					case cloudprovider.GCP:
-						cl := tc.tfClient.(*stubTerraformClient)
-						assert.True(cl.destroyClusterCalled)
-						assert.True(cl.cleanUpWorkspaceCalled)
-					case cloudprovider.Azure:
-						cl := tc.azureclient.(*stubAzureClient)
-						assert.True(cl.terminateResourceGroupResourcesCalled)
 					}
 				}
 			} else {