+++ cli/internal/cloudcmd/terminate.go
@@ -10,7 +10,6 @@ import (
 	"context"
 	"fmt"
 
-	azurecl "github.com/edgelesssys/constellation/v2/cli/internal/azure/client"
 	"github.com/edgelesssys/constellation/v2/cli/internal/libvirt"
 	"github.com/edgelesssys/constellation/v2/cli/internal/terraform"
 	"github.com/edgelesssys/constellation/v2/internal/cloud/cloudprovider"
@@ -19,19 +18,15 @@ import (
 
 // Terminator deletes cloud provider resources.
 type Terminator struct {
-	newTerraformClient func(ctx context.Context) (terraformClient, error)
-	newAzureClient     func(subscriptionID, tenantID string) (azureclient, error)
+	newTerraformClient func(ctx context.Context, provider cloudprovider.Provider) (terraformClient, error)
 	newLibvirtRunner   func() libvirtRunner
 }
 
 // NewTerminator create a new cloud terminator.
 func NewTerminator() *Terminator {
 	return &Terminator{
-		newTerraformClient: func(ctx context.Context) (terraformClient, error) {
-			return terraform.New(ctx, cloudprovider.GCP)
-		},
-		newAzureClient: func(subscriptionID, tenantID string) (azureclient, error) {
-			return azurecl.NewFromDefault(subscriptionID, tenantID)
+		newTerraformClient: func(ctx context.Context, provider cloudprovider.Provider) (terraformClient, error) {
+			return terraform.New(ctx, provider)
 		},
 		newLibvirtRunner: func() libvirtRunner {
 			return libvirt.New()
@@ -42,37 +37,22 @@ func NewTerminator() *Terminator {
 // Terminate deletes the could provider resources defined in the constellation state.
 func (t *Terminator) Terminate(ctx context.Context, state state.ConstellationState) error {
 	provider := cloudprovider.FromString(state.CloudProvider)
-	switch provider {
-	case cloudprovider.Azure:
-		cl, err := t.newAzureClient(state.AzureSubscription, state.AzureTenant)
-		if err != nil {
-			return err
-		}
-		return t.terminateAzure(ctx, cl, state)
-	case cloudprovider.GCP:
-		cl, err := t.newTerraformClient(ctx)
-		if err != nil {
-			return err
-		}
-		defer cl.RemoveInstaller()
-		return t.terminateTerraform(ctx, cl)
-	case cloudprovider.QEMU:
-		cl, err := t.newTerraformClient(ctx)
-		if err != nil {
-			return err
-		}
-		defer cl.RemoveInstaller()
+	if provider == cloudprovider.Unknown {
+		return fmt.Errorf("unknown cloud provider %s", state.CloudProvider)
+	}
+
+	cl, err := t.newTerraformClient(ctx, provider)
+	if err != nil {
+		return err
+	}
+	defer cl.RemoveInstaller()
+
+	if provider == cloudprovider.QEMU {
 		libvirt := t.newLibvirtRunner()
 		return t.terminateQEMU(ctx, cl, libvirt)
-	default:
-		return fmt.Errorf("unsupported provider: %s", provider)
 	}
-}
 
-func (t *Terminator) terminateAzure(ctx context.Context, cl azureclient, state state.ConstellationState) error {
-	cl.SetState(state)
-
-	return cl.TerminateResourceGroupResources(ctx)
+	return t.terminateTerraform(ctx, cl)
 }
 
 func (t *Terminator) terminateTerraform(ctx context.Context, cl terraformClient) error {