+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 <!-- For changes in existing functionality.  -->
 - Verify measurements using [Rekor](https://github.com/sigstore/rekor) transparency log.
+- The `constellation create` on Azure now uses Terraform to create and destroy cloud resources.
 
 ### Deprecated
 <!-- For soon-to-be removed features. -->